refactor(transport-primitives): adopt PeerBlobClient + TofuTrustStore
File moves:
  rust/kernel/src/hal/peer.rs          -> rust/transport-primitives/src/peer_blob_client.rs
  rust/raft/src/federation/tofu.rs     -> rust/transport-primitives/src/tofu.rs

Both lift to the lowest shared layer in the workspace dep graph so
raft (server-side fetcher) and rpc (client-side fetcher, Phase E) can
both name them without depending on each other:

  PeerBlobClient — transport-layer abstraction. Reference impl moves
  to rpc/peer_blob.rs in Phase E.2; no-op fallback ships here for
  non-cdylib builds.

  TofuTrustStore — peer zone CA fingerprint pinning (SSH-style).
  Currently consumed by raft federation via the join token verification
  path; rpc-side consumes it directly when peer-blob client moves
  out of transport.

Re-exports preserve existing import paths:
  - kernel::hal::peer::PeerBlobClient stays as a re-export module
  - nexus_raft::federation::TofuTrustStore re-exports the shared type
  - transport-primitives gains a `python` feature toggle so the cdylib
    pulls PyTofuTrustStore / PyTrustedZone pyclasses through raft's
    `python` feature passthrough (no API change for callers).

transport-primitives picks up serde, serde_json, time, pem, base64
runtime deps and rcgen / tempfile dev-deps so the relocated TOFU
unit tests keep working.

cargo check + clippy + fmt clean. Codegen refreshed.
