+++ rust/raft/src/pyo3_bindings.rs
@@ -826,7 +826,7 @@ pub fn register_python_classes(m: &Bound<'_, PyModule>) -> PyResult<()> {
 
     #[cfg(feature = "grpc")]
     {
-        use crate::federation::tofu::{PyTofuTrustStore, PyTrustedZone};
+        use transport_primitives::{PyTofuTrustStore, PyTrustedZone};
         m.add_class::<PyTofuTrustStore>()?;
         m.add_class::<PyTrustedZone>()?;
     }
