+++ rust/transport-primitives/src/lib.rs
@@ -1,38 +1,29 @@
-//! Shared gRPC transport primitives for the Nexus workspace.
+//! Shared transport primitives for the Nexus workspace — pure
+//! utilities consumed by both raft (server-side) and rpc (client-side)
+//! drivers, plus the kernel-bound front-door transport crate.
 //!
-//! Provides TLS configuration, peer addressing, connection pooling, and
-//! channel creation utilities. Consumed by `nexus_runtime` (ObjectStore gRPC
-//! adapter) and `nexus_raft` (Raft transport).
-//!
-//! This crate unifies tonic across the workspace (0.13) and eliminates
-//! duplicated transport boilerplate.
-//!
-//! ## Phase 4 note
-//!
-//! Per `KERNEL-ARCHITECTURE.md`, the kernel-bound high-level transport
-//! modules (gRPC server, IPC envelope, federation client) **were
-//! intended to live here** as `transport::{grpc, ipc, federation}`.
-//! They didn't move out of kernel in this Phase 4 pass because the
-//! pre-existing `raft -> transport` edge plus the planned
-//! `transport -> kernel` edge plus the existing `kernel -> raft` edge
-//! would close a Cargo dependency cycle.  Resolving that cleanly
-//! requires splitting this crate into `transport-primitives` (raft
-//! consumes) + `transport` (kernel-bound) — a follow-up surgery
-//! whose churn would dwarf the rest of Phase 4.
-//!
-//! Today the high-level modules live at `kernel::transport::*` (Linux
-//! precedent — networking lives in the kernel).  When the
-//! transport-primitives split lands, those files migrate here without
-//! source changes (only the crate name in their `use` paths).
+//! Provides TLS configuration, peer addressing, connection pooling,
+//! channel creation, the `PeerBlobClient` trait, and the TOFU
+//! (Trust On First Use) trust store. The crate sits at the lowest
+//! shared layer in the workspace dep graph: zero peer-crate deps,
+//! every higher-layer crate (raft, rpc, transport, kernel) depends
+//! on this one.
 
 mod channel;
 mod config;
 mod error;
 mod peer;
+mod peer_blob_client;
 mod pool;
+mod tofu;
 
 pub use channel::create_channel;
 pub use config::{ClientConfig, ServerConfig, TlsConfig};
 pub use error::{Result, TransportError};
 pub use peer::{compute_node_id, hostname_to_node_id, NodeAddress, PeerAddress};
+pub use peer_blob_client::{NoopPeerBlobClient, PeerBlobClient, PeerBlobResult};
 pub use pool::ConnectionPool;
+pub use tofu::{TofuError, TofuResult, TofuTrustStore, TrustedZone};
+
+#[cfg(feature = "python")]
+pub use tofu::{PyTofuTrustStore, PyTrustedZone};
